feat: refatorações de int to double para price e media de preço de compra

+++ app/src/main/kotlin/org/example/application/usecase/GetDepth.kt
@@ -27,5 +27,5 @@ data class GetDepthOutput(
     val sells: MutableList<Order>
 ) {
     @Serializable
-    data class Order(val quantity: Int, val price: Int)
+    data class Order(val quantity: Int, val price: Double)
 }
